+++ pkg/sentry/kernel/callbacks.go
@@ -2,20 +2,29 @@ package kernel
 
 import (
 	"errors"
+	"fmt"
+	"github.com/dop251/goja"
 	"gvisor.dev/gvisor/pkg/sentry/arch"
+	"gvisor.dev/gvisor/pkg/sentry/kernel/callbacks"
+	"strings"
 	"sync"
 )
+
+type SyscallArgumentSubstitutionBeforeExecution struct {
+	rval uintptr
+	err  error
+}
 
 // CallbackBefore - interface which is used to observe and / or modify syscall arguments
 type CallbackBefore interface {
 	// CallbackBeforeFunc accepts Task, sysno and syscall arguments returns:
 	//
-	// TODO: return new args, new rval, new err, instead, error
-	CallbackBeforeFunc(t *Task, sysno uintptr, args *arch.SyscallArguments) (*arch.SyscallArguments, error)
+	// new args, rval/err if needed, error if something bad occurred
+	CallbackBeforeFunc(t *Task, sysno uintptr, args *arch.SyscallArguments) (*arch.SyscallArguments, *SyscallArgumentSubstitutionBeforeExecution, error)
 }
 
 type CallbackAfter interface {
-	// CallbackAfterFunc accepts Task, sysno and syscall arguments
+	// CallbackAfterFunc accepts Task, sysno, syscall arguments and rval, err after as result of gvisor syscall impl
 	//
 	// - new args
 	//
@@ -24,7 +33,7 @@ type CallbackAfter interface {
 	// - new err (should be converted to golang error)
 	//
 	// - error if something went wrong
-	CallbackAfterFunc(t *Task, sysno uintptr, args *arch.SyscallArguments) (*arch.SyscallArguments, uintptr, uintptr, error)
+	CallbackAfterFunc(t *Task, sysno uintptr, args *arch.SyscallArguments, rval uintptr, err error) (*arch.SyscallArguments, uintptr, uintptr, error)
 }
 
 type CallbackTable struct {
@@ -141,4 +150,86 @@ func (s *SimplePrinter) CallbackBeforeFunc(t *Task, sysno uintptr, args *arch.Sy
 	s.mu.Unlock()
 	t.Debugf("sysno %v: Bruh... %v", sysno, val)
 	return args, nil
+}
+
+// JsCallbackBefore implements CallbackBefore
+type JsCallbackBefore struct {
+	source     string
+	entryPoint string
+	sysno      uintptr
+}
+
+type JsCallbackAfter struct {
+}
+
+func (cb *JsCallbackBefore) fromDto(dto *callbacks.CallbackDto) error {
+	if dto.EntryPoint == "" || dto.CallbackSource == "" || dto.Type != "before" {
+		return errors.New("invalid before callback dto")
+	}
+
+	cb.sysno = uintptr(dto.Sysno)
+	cb.source = dto.CallbackSource
+	cb.entryPoint = dto.EntryPoint
+	return nil
+}
+
+// addSyscallArgsToContextObject from this context object user`s callback will take syscall args
+func addSyscallArgsToContextObject(object *goja.Object, arguments *arch.SyscallArguments) error {
+	for i, arg := range arguments {
+		err := object.Set(fmt.Sprintf("arg%d", i), int64(arg.Value))
+
+		if err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
+// callbackInvocationTemplate generate string that represent user callback script + invocation of it with injected args
+func (cb *JsCallbackBefore) callbackInvocationTemplate() string {
+	args := make([]string, len(arch.SyscallArguments{}))
+	for i := range args {
+		args[i] = fmt.Sprintf("args.arg%d", i)
+	}
+
+	return fmt.Sprintf("%s; %s(%s)", cb.source, cb.entryPoint, strings.Join(args, ", "))
+}
+
+// CallbackBeforeFunc execution of user callback for syscall on js VM with our hooks
+func (cb *JsCallbackBefore) CallbackBeforeFunc(t *Task, _ uintptr, args *arch.SyscallArguments) (*arch.SyscallArguments, error) {
+	kernel := t.Kernel()
+	kernel.GojaRuntime.Mutex.Lock()
+	defer kernel.GojaRuntime.Mutex.Unlock()
+
+	vm := kernel.GojaRuntime.JsVM
+	hooksHolder := vm.NewObject()
+	if err := kernel.hooksTable.addHooksToContextObject(hooksHolder, t); err != nil {
+		return nil, err
+	}
+
+	if err := vm.Set("hooks", hooksHolder); err != nil {
+		return nil, err
+	}
+
+	argsHolder := vm.NewObject()
+	if err := addSyscallArgsToContextObject(argsHolder, args); err != nil {
+		return nil, err
+	}
+
+	if err := vm.Set("args", argsHolder); err != nil {
+		return nil, err
+	}
+
+	val, err := vm.RunString(cb.callbackInvocationTemplate())
+	if err != nil {
+		return nil, err
+	}
+
+	ret, err_ := callbacks.ExtractArgsFromRetJsValue(args, vm, &val)
+	if err_ != nil {
+		return nil, err_
+	}
+
+	return ret, nil
 }
